Extract chuck norris component
